Add unit tests, fix psalm errors, fix phpcs violations
Tests (47 tests, 102 assertions):
- HeapStatsTest: parseSize, humanReadableBytes, formatSize
- CooldownManagerTest: cooldown, exponential backoff, backoff cap,
  clear reset, hourly rate limit, max triggers, independent triggers,
  skip reason, total fires
- DiskUsageTrackerTest: within limit, exceeds limit, unlimited, bytes
- TriggerEventTest: construction, null value
- MemoryLimitTriggerTest: above/below/at limit, properties
- MemoryGrowthRateTriggerTest: parseRate, no previous, exceeds/below
  rate, memory decrease, properties
- MemoryPeakTriggerTest: no previous, peak changed/unchanged
- FunctionDetectionTriggerTest: found, class method, not found, no trace
- TraceDepthTriggerTest: exceeded, within limit, no trace
- ExceptionDetectionTriggerTest: in flight, no exception, properties
- LogActionTest: output format, toFile

Psalm fixes (14 errors → 0):
- Cast float/int operations explicitly per strict mode
- Add @var/@param type annotations for version strings
- Move CooldownState to own file (PSR-12)
- Fix SplQueue generic type annotation

Phpcs fixes (PSR-12):
- Break long lines (>120 chars) across multiple lines
- One class per file (CooldownState extracted)

+++ src/Inspector/Watch/HeapStats.php
@@ -48,9 +48,9 @@ final class HeapStats
         $last = strtoupper(substr($size, -1));
         $value = (float)substr($size, 0, -1);
         return match ($last) {
-            'G' => (int)($value * 1024 * 1024 * 1024),
-            'M' => (int)($value * 1024 * 1024),
-            'K' => (int)($value * 1024),
+            'G' => (int)((int)$value * 1024 * 1024 * 1024),
+            'M' => (int)((int)$value * 1024 * 1024),
+            'K' => (int)((int)$value * 1024),
             default => (int)$size,
         };
     }
